% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Med_Effs.R
\name{all_cov_MEs}
\alias{all_cov_MEs}
\title{Covariance matrix of all mediation effects on various scales.}
\usage{
all_cov_MEs(scale = c("diff", "rat", "OR"), w, fit_Y, fit_M)
}
\arguments{
\item{scale}{The scale(s) of the mediation effect. Can be "diff", "rat" or "OR".}

\item{w}{Level of covariates, \eqn{W}.}

\item{fit_Y, fit_M}{Fitted models for Y and M.}
}
\value{
A covariance matrix for all mediation effects (total, direct and indirect) on the specified scale(s).
}
\description{
Covariance matrix of all mediation effects on various scales.
}
\details{
Note: Uses the \eqn{K}-adjusted covariance matrix, not the asymptotic covariance matrix.
}
