% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Med_Effs.R
\name{all_grad_MEs}
\alias{all_grad_MEs}
\alias{all_grad_MEs_pars}
\alias{all_grad_MEs_models}
\title{Jacobian of all mediation effects wrt all ENCs}
\usage{
all_grad_MEs_pars(
  scale = c("diff", "rat", "OR"),
  w,
  b_Y,
  theta_Y,
  b_M,
  theta_M
)

all_grad_MEs_models(scale, w, fit_Y, fit_M)
}
\arguments{
\item{scale}{The scale(s) of the mediation effect. Can be "diff", "rat" or "OR".}

\item{w}{Level of covariates, \eqn{W}.}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}
}
\value{
A matrix of partial derivatives. Dimension is (3 * length(scale))-by-4.
}
\description{
Jacobian of all mediation effects wrt all ENCs
}
