% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_Par_Covs.R
\name{all_pars_cov_mat}
\alias{all_pars_cov_mat}
\title{Joint (sample-size adjusted) covariance matrix of fixed and random effects' parameters from both models}
\usage{
all_pars_cov_mat(fit_Y, fit_M)
}
\arguments{
\item{fit_Y}{Model for \code{Y}}

\item{fit_M}{Model for \code{M}}
}
\value{
A matrix of the joint covariance matrix of fixed and random effects' parameters from both models. Obtained from asymptotic theory, then re-scaled to match the observed number of groups.
}
\description{
Joint (sample-size adjusted) covariance matrix of fixed and random effects' parameters from both models
}
\details{
Note: We assume that parameter estimators from the two models are independent. This may not be true in practice. See, e.g., Bauer, Preacher, & Gil (2006) for a method to incorporate inter-model dependence.
}
