% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_Helpers.r
\name{all_pars_cov_mat_TMB}
\alias{all_pars_cov_mat_TMB}
\title{Estimated sampling covariance matrix for all GLMM parameters (i.e. from both models)}
\usage{
all_pars_cov_mat_TMB(fit_Y, fit_M, return_separate_covs = FALSE)
}
\arguments{
\item{fit_Y}{A GLMM fit using \code{glmmTMB} for the outcome model}

\item{fit_M}{A GLMM fit using \code{glmmTMB} for the mediator model}
}
\value{
Estimated sampling covariance matrix for all parameters from both models following our parameterization.
}
\description{
Estimated sampling covariance matrix for all GLMM parameters (i.e. from both models)
}
