% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_Functions.r
\name{CI_Widths}
\alias{CI_Widths}
\alias{get_widths_one_cov_mat}
\alias{get_widths_many_cov_mats}
\title{Construct CIs and compute their widths}
\usage{
get_widths_one_cov_mat(Theta_hats_data, cov_mat)

get_widths_many_cov_mats(Theta_hats_data, cov_mat_list)
}
\arguments{
\item{Theta_hats_data}{A data frame containing estimates of each parameter. Each column is a parameter, and each row is a dataset.}

\item{cov_mat}{An estimated sampling covariance matrix for the estimates.}

\item{cov_mat_list}{A list of estimated sampling covariance matrices, one for each dataset.}
}
\value{
A list of vectors containing widths of all intervals for each parameter. The list has one component per parameter, and the vector has one component per dataset.
}
\description{
Construct CIs and compute their widths
}
