% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Average_Mediation_Effects.r
\name{data_2_confounder_dist}
\alias{data_2_confounder_dist}
\title{Extract the distribution of confounders from a dataset}
\usage{
data_2_confounder_dist(
  data,
  outcome_name,
  exposure_name,
  mediator_name,
  group_name
)
}
\arguments{
\item{data}{A data frame containing mediation variables and confounders}

\item{outcome_name, exposure_name, mediator_name, group_name}{Names of the outcome, exposure, mediator, and group variables respectively. Must be strings.}
}
\value{
A list with two components: `values`, a list of vectors of confounder values, and `freqs`, a vector of frequencies of each confounder value.
}
\description{
Extract the distribution of confounders from a dataset
}
