% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R
\name{ENC_cross_covariances}
\alias{ENC_cross_covariances}
\alias{cross_cov_ENC}
\alias{cross_cov_ENC_Sigma}
\alias{cross_cov_ENC_pars}
\title{Cross-Covariances for ENCs}
\usage{
cross_cov_ENC(
  w1,
  w2,
  fit_Y,
  fit_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

cross_cov_ENC_Sigma(
  w1,
  w2,
  Sigma,
  fit_Y,
  fit_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

cross_cov_ENC_pars(
  w1,
  w2,
  Sigma,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)
}
\arguments{
\item{w1, w2}{Two vectors of confounders}

\item{fit_Y, fit_M}{Fitted models for Y and M.}

\item{which_REs}{Which random effects to include in the calculation. Default is all. Shorthands are available. See details.}

\item{Sigma}{Covariance matrix of the model parameters.}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}
}
\value{
The cross-covariance matrix of all ENC estimates evaluated at two levels of the confounders. Order of \eqn{(X, X_M)} levels is (1,1), (1,0), (0,1), (0,0).
}
\description{
Cross-Covariances for ENCs
}
\details{
Note: Uses the \eqn{K}-adjusted covariance matrix, not the asymptotic covariance matrix.

Note: Setting `w1==w2` will give the covariance matrix of the ENC estimates.

The following shorthands for random effects are available:
\itemize{
\item "All": All REs
\item "Y.All": All REs for Y
\item "M.All": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
