% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Helpers.R
\name{expand_REs}
\alias{expand_REs}
\title{Expand shorthand notation for random effects in models for Y and M}
\usage{
expand_REs(RE_input)
}
\arguments{
\item{RE_input}{A character vector of REs to include. May contain shorthands. See details.}
}
\value{
A character vector containing names of the REs to include in the models.
}
\description{
Expand shorthand notation for random effects in models for Y and M
}
\details{
The following shorthands for random effects are available:
\itemize{
\item "all": All REs
\item "Y.all": All REs for Y
\item "M.all": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
