% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R
\name{gamma_vecs}
\alias{gamma_vecs}
\alias{Y_vec_gamma}
\alias{M_vec_gamma}
\title{Vectors used to compute gamma}
\usage{
Y_vec_gamma(x, m, which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X"))

M_vec_gamma(x, which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X"))
}
\arguments{
\item{x, m}{Values of X and M at which gamma is evaluated.}

\item{which_REs}{Which random effects to include in the calculation. Default is all. Shorthands are available. See details.}
}
\value{
A vector to be used in the quadratic form defining gamma.
}
\description{
Vectors used to compute gamma
}
\details{
The following shorthands for random effects are available:
\itemize{
\item "all": All REs
\item "Y.all": All REs for Y
\item "M.all": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
