% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_Functions.r
\name{get_coverage_rates}
\alias{get_coverage_rates}
\title{Compute coverage rates for Wald CIs using a single covariance matrix for each estimate}
\usage{
get_coverage_rates(Theta_hats_data, cov_mat, true_Thetas)
}
\arguments{
\item{Theta_hats_data}{A data frame containing one estimated parameter per column and one dataset per row}

\item{cov_mat}{A covariance matrix to use with each estimate}

\item{true_Thetas}{True values of each parameter for checking coverage}
}
\value{
A vector of coverage rates, one component for each parameter
}
\description{
Compute coverage rates for Wald CIs using a single covariance matrix for each estimate
}
