% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_Functions.r
\name{get_coverage_rates_many_cov_mats}
\alias{get_coverage_rates_many_cov_mats}
\title{Compute coverage rates for Wald CIs using a separate covariance matrix for each estimate}
\usage{
get_coverage_rates_many_cov_mats(Theta_hats_data, cov_mat_list, true_Thetas)
}
\arguments{
\item{Theta_hats_data}{A data frame containing one estimated parameter per column and one dataset per row}

\item{cov_mat_list}{A list of covariance matrices, one for each dataset}

\item{true_Thetas}{True values of each parameter for checking coverage}
}
\value{
A vector of coverage rates, one component for each parameter
}
\description{
Compute coverage rates for Wald CIs using a separate covariance matrix for each estimate
}
