% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Sample_Data.R
\name{get_lin_preds}
\alias{get_lin_preds}
\title{Compute the linear predictor based on provided datasets and parameters}
\usage{
get_lin_preds(
  data_fix,
  data_ran,
  beta,
  Gamma,
  add_intercept = NULL,
  add_intercept_fix = NULL,
  add_intercept_ran = NULL,
  return_REs = FALSE
)
}
\arguments{
\item{data_fix}{A dataset containing covariates with fixed effects. Intercept can optionally be added later.}

\item{data_ran}{A dataset containing coveriates with random effects. Intercept can optionally be added later.}

\item{beta}{Vector of fixed effects coefficients.}

\item{Gamma}{Covariance matrix of the random effects coefficients.}

\item{add_intercept}{Should an intercept column be added to the datasets for both fixed and random effects?}

\item{add_intercept_fix, add_intercept_ran}{Should an intercept column be added to the datasets for fixed and random effects respectively?}

\item{return_REs}{Should generated random effects be returned?}
}
\value{
A vector containing a linear predictor for each observation in the provided datasets.
}
\description{
Note : Please specify either add_intercept or both add_intercept_fix and add_intercept_ran. Do not provide all three!
}
\examples{
data_fix = data.frame(X1 = c(1,0), X2 = c(0,1))
data_ran = data_fix

# Fixed and random effects for intercept
beta1 = c(1,2,3)
Gamma1 = diag(3)

get_lin_preds(data_fix, data_ran, beta1, Gamma1,
  add_intercept = TRUE)


# Fixed effect only for intercept
beta2 = c(1,2,3)
Gamma2 = diag(2)

get_lin_preds(data_fix, data_ran, beta2, Gamma2,
  add_intercept_fix = TRUE, add_intercept_ran = FALSE)


# Return generated random effects in the second example
get_lin_preds(data_fix, data_ran, beta2, Gamma2,
 add_intercept_fix = TRUE, add_intercept_ran = FALSE, return_REs = TRUE)
}
