% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Med_Effs.R
\name{get_ME}
\alias{get_ME}
\title{Compute a mediation effect on the given scale(s)}
\usage{
get_ME(ENC1, ENC2, scale = c("diff", "rat", "OR"))
}
\arguments{
\item{ENC1, ENC2}{Two expected nested counterfactuals}

\item{scale}{The scale(s) of the mediation effect. Can be "diff", "rat" or "OR".}
}
\value{
The mediation effect on the given scale.
}
\description{
Compute a mediation effect on the given scale(s)
}
