% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Helpers.R
\name{get_model_pars}
\alias{get_model_pars}
\title{Extract fitted parameters of interest from an lme4 model.}
\usage{
get_model_pars(fit, format = "list")
}
\arguments{
\item{fit}{An lme4 model}

\item{format}{The format of the output. Can be "list" or "vector".}
}
\value{
A list with elements b (fixed effects) and theta (RE SDs and correlations), or a vector with b followed by theta.
}
\description{
Extract fitted parameters of interest from an lme4 model.
}
