% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Helpers.R
\name{get_model_pars_TMB}
\alias{get_model_pars_TMB}
\title{Extract the parameters (in our order) from a glmmTMB object}
\usage{
get_model_pars_TMB(fit, format = "list")
}
\arguments{
\item{fit}{A GLMM fit using glmmTMB}

\item{format}{The format of the output. Can be "list" or "vector".}
}
\value{
A vector, theta, of model parameters. Order is fixed effects, then covariance parameters. The latter is organized as, e.g., SD, corr, corr, SD, corr, SD for a 3x3 covariance matrix.
}
\description{
Extract the parameters (in our order) from a glmmTMB object
}
