% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_Functions.r
\name{get_widths_one_par}
\alias{get_widths_one_par}
\title{Compute CI widths from endpoints}
\usage{
get_widths_one_par(CIs_one_par)
}
\arguments{
\item{CIs_one_par}{A list with two elements: `lcl` and `ucl`. Each component is a vector of CI endpoints.}
}
\value{
A vector of CI widths
}
\description{
Compute CI widths from endpoints
}
