% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{grad_psi}
\alias{grad_psi}
\alias{d1_psi}
\alias{d2_psi}
\title{Gradient of the intermediate integral function, psi}
\usage{
d1_psi(mu, sigma)

d2_psi(mu, sigma)
}
\arguments{
\item{mu}{Mean of the latent Gaussian.}

\item{sigma}{SD of the latent Gaussian.}
}
\value{
The gradient of the function \code{psi(mu, sigma)} with respect to \code{mu} and \code{sigma}.
}
\description{
Gradient of the intermediate integral function, psi
}
