% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{grad_psi_chain}
\alias{grad_psi_chain}
\title{Gradient of psi, chain rule version}
\usage{
grad_psi_chain(mu, sigma, grad_mu, grad_sigma)
}
\arguments{
\item{mu, sigma}{Arguments of psi}

\item{grad_mu, grad_sigma}{Gradients of the arguments, \code{mu} and \code{sigma}.}
}
\value{
Total gradient of psi.
}
\description{
This function computes the gradient of the function \code{psi(mu, sigma)}, where \code{mu} and \code{sigma} are themselves functions of the variable with respect to which we are differentiating.
}
