% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_Helpers.r
\name{grad_TMB_2_GLMM}
\alias{grad_TMB_2_GLMM}
\alias{TMB_2_GLMM_grad_FE}
\alias{TMB_2_GLMM_grad_SD}
\alias{TMB_2_GLMM_grad_corr}
\alias{TMB_2_GLMM_grad}
\title{Gradient of GLMM parameters under our parameterization WRT glmmTMB parameterization.}
\usage{
TMB_2_GLMM_grad_FE(fit_TMB)

TMB_2_GLMM_grad_SD(fit_TMB)

TMB_2_GLMM_grad_corr(fit_TMB)

TMB_2_GLMM_grad(fit_TMB)
}
\arguments{
\item{fit_TMB}{A model fit using \code{glmmTMB}}
}
\value{
A matrix of gradients. Rows index outputs and columns index inputs.
}
\description{
Gradient of GLMM parameters under our parameterization WRT glmmTMB parameterization.
}
\details{
`TMB_2_GLMM_grad_FE()` Returns the gradients of the fixed effects parameters (WRT all glmmTMB parameters) \
`TMB_2_GLMM_grad_SD()` Returns the gradients of the standard deviations (WRT all glmmTMB parameters) \
`TMB_2_GLMM_grad_FE()` Returns the gradients of the correlations (WRT all glmmTMB parameters)

`TMB_2_GLMM_grad()` Combines the three sets of gradients and permutes them to match our parameterization.
}
