% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R
\name{Jacob_ENC}
\alias{Jacob_ENC}
\alias{Jacob_ENC_pars}
\alias{Jacob_ENC_models}
\alias{Jacob_ENC_Theta}
\title{"Jacobian" of expected nested counterfactuals. I.e. Gradients for every combination of \eqn{X, X_M}.}
\usage{
Jacob_ENC_pars(
  w,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

Jacob_ENC_models(
  w,
  fit_Y,
  fit_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

Jacob_ENC_Theta(
  w,
  Theta,
  len_par_vecs,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)
}
\arguments{
\item{w}{Level of covariates, \eqn{W}.}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}

\item{which_REs}{Which random effects to include in the calculation. Default is all. Shorthands are available. See details.}

\item{fit_Y, fit_M}{Fitted models for Y and M.}

\item{Theta}{Vector of parameters from both models. Order is b_Y, theta_Y, b_M, theta_M}
}
\value{
Gradients for every combination of \eqn{X} and \eqn{X_M}, organized as a 4-by-many matrix. Order of \eqn{(X, X_M)} levels is (1,1), (1,0), (0,1), (0,0). Length of each gradient (i.e. number of columns) is adjusted to match which_REs.
}
\description{
"Jacobian" of expected nested counterfactuals. I.e. Gradients for every combination of \eqn{X, X_M}.
}
\details{
The following shorthands for random effects are available:
\itemize{
\item "All": All REs
\item "Y.All": All REs for Y
\item "M.All": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
