% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Sample_Data.R
\name{list_2_data_make_labels}
\alias{list_2_data_make_labels}
\title{Create group labels}
\usage{
list_2_data_make_labels(X_list, group_labels = NULL)
}
\arguments{
\item{X_list}{A list of data frames.}

\item{group_labels}{Optionally, a vector of labels for each element of `X_list`. Must be either the same length as `X_list` or equal to the total number of rows among elements in `X_list`. If `NULL`, labels are `G1`, `G2`,...}
}
\value{
A vector of group labels with length equal to the total number of rows among elements of `X_list`.
}
\description{
Create group labels
}
\examples{
data = as.data.frame(matrix(c(1,0,0,1), nrow = 2))
more_data = as.data.frame(matrix(c(0,1,1,0), nrow = 2))
data_list = list(data, more_data)

list_2_data_make_labels(data_list)

data_names = c("Alice", "Bob")
list_2_data_make_labels(data_list, data_names)
}
\keyword{internal}
