% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericHelpers.R
\name{logit_expit}
\alias{logit_expit}
\alias{logit}
\alias{expit}
\title{Title}
\usage{
logit(x)

expit(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector of the same length as x
}
\description{
Title
}
\examples{
logit(c(0.1, 0.2, 0.3))
expit(c(0.1, 0.2, 0.3))

logit(expit(c(0.1, 0.2, 0.3))
expit(logit(c(0.1, 0.2, 0.3))
}
