% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Bootstrap_Data.R,
%   R/Generate_Sample_Data.R
\name{make_all_reg_pars}
\alias{make_all_reg_pars}
\title{Store all regression parameters in a single list}
\usage{
make_all_reg_pars(
  beta_Y = c(1, -1.5, 1, -1.5, 1),
  Gamma_Y = diag(3),
  beta_M = c(-1, 1, -1, 1),
  Gamma_M = diag(2)
)

make_all_reg_pars(
  beta_Y = c(1, -1.5, 1, -1.5, 1),
  Gamma_Y = diag(3),
  beta_M = c(-1, 1, -1, 1),
  Gamma_M = diag(2)
)
}
\arguments{
\item{beta_Y}{Vector of length 4 containing fixed-effects for the outcome model. Order of variables is: Intercept, M, X, C1, C2.}

\item{Gamma_Y}{Covariance matrix of size 3x3 containing random effects for the outcome model. Order of variables is: Intercept, M, X.}

\item{beta_M}{Vector of length 3 containing fixed-effects for the outcome model. Order is: Intercept, X, C1, C2.}

\item{Gamma_M}{Covariance matrix of size 2x2 containing random effects for the outcome model. Order of variables is: Intercept, X.}
}
\value{
A named list containing fixed and random effects parameters for both the outcome and mediator regression models. Names in list match names of arguments to this function.

A named list containing fixed and random effects parameters for both the outcome and mediator regression models. Names in list match names of arguments to this function.
}
\description{
Store all regression parameters in a single list

Store all regression parameters in a single list
}
\examples{
make_all_reg_pars()
make_all_reg_pars()
}
