% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_Functions.r
\name{many_widths}
\alias{many_widths}
\title{Compute CI widths from endpoints for multiple parameters}
\usage{
many_widths(CIs_many_pars)
}
\arguments{
\item{CIs_many_pars}{A list of lists. Each sub-list contains two elements: `lcl` and `ucl`. Each of `lcl` and `ucl` are vectors of CI endpoints.}
}
\value{
A list of vectors of CI widths. Each vector contains the widths of all intervals for one parameter
}
\description{
Compute CI widths from endpoints for multiple parameters
}
