% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Med_Effs.R
\name{ME_scales}
\alias{ME_scales}
\alias{ME_diff}
\alias{ME_rat}
\alias{ME_OR}
\title{Compute mediation effects on different scales}
\usage{
ME_diff(ENC1, ENC2)

ME_rat(ENC1, ENC2)

ME_OR(ENC1, ENC2)
}
\arguments{
\item{ENC1, ENC2}{Two expected nested counterfactuals}
}
\value{
The mediation effect on the given scale.
}
\description{
Compute mediation effects on different scales
}
