% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_Functions.r
\name{Mean_CI_Widths}
\alias{Mean_CI_Widths}
\alias{mean_widths_one_cov_mat}
\alias{mean_widths_many_cov_mats}
\title{Mean CI width for each parameter}
\usage{
mean_widths_one_cov_mat(Theta_hats_data, cov_mat)

mean_widths_many_cov_mats(Theta_hats_data, cov_mat_list)
}
\arguments{
\item{Theta_hats_data}{A data frame containing estimates of each parameter. Each column is a parameter, and each row is a dataset.}

\item{cov_mat}{An estimated sampling covariance matrix for the estimates.}

\item{cov_mat_list}{A list of estimated sampling covariance matrices, one for each dataset.}
}
\value{
A vector of mean CI widths, one component for each parameter.
}
\description{
Mean CI width for each parameter
}
