% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Med_Effs.R
\name{Med_Effs}
\alias{Med_Effs}
\alias{total_effect}
\alias{direct_effect}
\alias{indirect_effect}
\title{Mediation effects of \eqn{X} on \eqn{Y} mediated by \eqn{M}}
\usage{
total_effect(scale = c("diff", "rat", "OR"), w, b_Y, theta_Y, b_M, theta_M)

direct_effect(
  scale = c("diff", "rat", "OR"),
  w,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  x_m_ref = 0
)

indirect_effect(
  scale = c("diff", "rat", "OR"),
  w,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  x_ref = 1
)
}
\arguments{
\item{scale}{The scale(s) of the mediation effect. Can be "diff", "rat" or "OR".}

\item{w}{Level of covariates, \eqn{W}.}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}

\item{x_m_ref}{Reference value of \eqn{X} for determining the mediator level.}

\item{x_ref}{Reference value of \eqn{X}.}
}
\value{
The specified mediation effect of \eqn{X} on \eqn{Y} mediated by \eqn{M}.
}
\description{
Mediation effects of \eqn{X} on \eqn{Y} mediated by \eqn{M}
}
\details{
Contents of \code{b_Y} are \code{(b_Y_0, b_Y_X, b_Y_M, B_Y_W)}. Contents of \code{b_M} are \code{(b_M_0, b_M_X, B_M_W)}.

Contents of \code{theta_Y} are \code{(s_Y_0, cor_Y_0X, cor_Y_0M, s_Y_X, cor_Y_XM, s_Y_M)}. Contents of \code{theta_M} are \code{(s_M_0, cor_M_0X, s_M_X)}.
}
