% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMedUQ-package.R
\docType{package}
\name{MultiMedUQ-package}
\alias{MultiMedUQ}
\alias{MultiMedUQ-package}
\title{MultiMedUQ: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wruth1/MultiMedUQ}
  \item Report bugs at \url{https://github.com/wruth1/MultiMedUQ/issues}
}

}
\keyword{internal}
