% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Intermediate integral for computing mediation effects}
\usage{
psi(mu, sigma)
}
\arguments{
\item{mu}{Mean of the latent Gaussian.}

\item{sigma}{SD of the latent Gaussian.}
}
\value{
The integral of the function \code{dnorm(x)/(1 + exp(-mu - sigma*x))} from -Inf to Inf.
}
\description{
This function computes the integral of the function \code{dnorm(x)/(1 + exp(-mu - sigma*x))} from -Inf to Inf. I.e. The expected inverse-logit of a RV with mean \code{mu} and SD \code{sigma}.
This is useful for computing various mediation effects.
}
