% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_Functions.r
\name{SD_CI_Widths}
\alias{SD_CI_Widths}
\alias{SD_widths_one_cov_mat}
\alias{SD_widths_many_cov_mats}
\title{SD of CI widths for each parameter}
\usage{
SD_widths_one_cov_mat(Theta_hats_data, cov_mat)

SD_widths_many_cov_mats(Theta_hats_data, cov_mat_list)
}
\arguments{
\item{Theta_hats_data}{A data frame containing estimates of each parameter. Each column is a parameter, and each row is a dataset.}

\item{cov_mat}{An estimated sampling covariance matrix for the estimates.}

\item{cov_mat_list}{A list of estimated sampling covariance matrices, one for each dataset.}
}
\value{
A vector of standard deviations of CI widths, one component for each parameter.
}
\description{
SD of CI widths for each parameter
}
