% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Helpers.R
\name{sds_corrs2theta}
\alias{sds_corrs2theta}
\title{Re-arrange a vector of SDs followed by correlations to my theta order}
\usage{
sds_corrs2theta(sd_corr_vec)
}
\arguments{
\item{sd_corr_vec}{A vector of SDs followed by correlations}
}
\value{
A vector of SDs and correlations in my theta order. Order is column major; e.g., SD_1, corr_12, corr_13, SD_2, corr_23, SD_3 for a 3x3 covariance matrix.
}
\description{
Re-arrange a vector of SDs followed by correlations to my theta order
}
