% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Helpers.R
\name{theta2Sigma}
\alias{theta2Sigma}
\title{Convert a vector, theta, to its corresponding covariance matrix}
\usage{
theta2Sigma(theta)
}
\arguments{
\item{theta}{A vector of SDs and correlations. Order matches that of merDeriv (i.e. SD, cor, SD for a 2x2).}
}
\value{
Sigma, a covariance matrix
}
\description{
Convert a vector, theta, to its corresponding covariance matrix
}
