% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_Helpers.r
\name{TMB_2_GLMM_pars}
\alias{TMB_2_GLMM_pars}
\title{Extract GLMM parameters from a glmmTMB object}
\usage{
TMB_2_GLMM_pars(fit_TMB)
}
\arguments{
\item{fit_TMB}{A model fit using \code{glmmTMB}}
}
\value{
Theta, a vector of model parameters. Order is FEs for Y, RE parameters for Y, FEs for M, RE parameters for M. See, e.g., [Sigma2theta()], for details on the order of RE parameters.
}
\description{
Uses our parameterization of SDs and correlations. Output is a vector organized according to our usual parameterization (see, e.g., [Sigma2theta()])
}
