% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_Helpers.r
\name{TMB_2_GLMM_pars_list}
\alias{TMB_2_GLMM_pars_list}
\title{Extract GLMM parameters from a glmmTMB object}
\usage{
TMB_2_GLMM_pars_list(fit_TMB)
}
\arguments{
\item{fit_TMB}{A model fit using \code{glmmTMB}}
}
\value{
A list containing fixed effects (\code{TMB_FEs}), standard deviations (\code{TMB_SDs}), and correlations (\code{TMB_corrs}).
}
\description{
Uses our parameterization of SDs and correlations. Mostly serves to translate from glmmTMB parameterization.
}
