% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_Helpers.r
\name{TMB_pars_list}
\alias{TMB_pars_list}
\title{Extract fitted internal parameters from a glmmTMB object}
\usage{
TMB_pars_list(fit_TMB)
}
\arguments{
\item{fit_TMB}{A model fit using \code{glmmTMB}}
}
\value{
A list containing fixed effects (\code{TMB_FE_pars}), log-standard deviations (\code{TMB_SD_pars}), and Cholesky factors from the (scaled) correlation matrix (\code{TMB_corr_pars}).
}
\description{
Uses the \code{glmmTMB} parameterization of log-SDs and Cholesky factors from the scaled correlation matrix (see the [TMB documentation](https://kaskr.github.io/adcomp/classdensity_1_1UNSTRUCTURED__CORR__t.html) for more details).
}
