% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_Helpers.r
\name{TMB_SE}
\alias{TMB_SE}
\alias{glmmTMB_SE_mat}
\alias{TMB_2_GLMM_SE}
\title{Covariance matrix for GLMM parameters from `glmmTMB`}
\usage{
glmmTMB_SE_mat(fit_TMB)

TMB_2_GLMM_SE(fit_TMB)
}
\arguments{
\item{fit_TMB}{A model fit using \code{glmmTMB}}
}
\value{
Estimated covariance matrix for the sampling distribution of GLMM parameters. `glmmTMB_SE_mat()` uses `glmmTMB`'s internal parameterization, while `TMB_2_GLMM_SE()` uses our parameterization.
}
\description{
Covariance matrix for GLMM parameters from `glmmTMB`
}
