% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot_eff.R
\name{tot_eff}
\alias{tot_eff}
\title{Total effect on odds-ratio scale}
\usage{
tot_eff(eta, zeta, a1, b1, b2, s1, s2, s3, s4)
}
\arguments{
\item{eta, zeta}{Linear predictors of M and Y models, respectively.}

\item{a1, b1, b2}{Coefficients of X in M, M in Y and X in Y, respectively.}

\item{s1, s2}{SD of linear predictor at x+1 in Y and M, respectively}

\item{s3, s4}{SD of linear predictor at x in Y and M, respectively}
}
\value{
The total effect of X on Y on the odds-ratio scale.
}
\description{
Total effect on odds-ratio scale
}
