% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R
\name{all_covs_ENC}
\alias{all_covs_ENC}
\title{Covariance matrix of all ENC estimates}
\usage{
all_covs_ENC(
  w,
  fit_Y,
  fit_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)
}
\arguments{
\item{w}{Level of covariates, \eqn{W}.}

\item{fit_Y, fit_M}{Fitted models for Y and M.}

\item{which_REs}{Which random effects to include in the calculation. Default is all. Shorthands are available. See details.}
}
\value{
The covariance matrix of all ENC estimates. Order of \eqn{(X, X_M)} levels is (1,1), (1,0), (0,1), (0,0).
}
\description{
Covariance matrix of all ENC estimates
}
\details{
Note: Uses the \eqn{K}-adjusted covariance matrix, not the asymptotic covariance matrix.

The following shorthands for random effects are available:
\itemize{
\item "All": All REs
\item "Y.All": All REs for Y
\item "M.All": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
