% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R
\name{all_covs_ENC}
\alias{all_covs_ENC}
\title{Covariance matrix of all ENC estimates}
\usage{
all_covs_ENC(w, fit_Y, fit_M)
}
\arguments{
\item{w}{Level of covariates, \eqn{W}.}

\item{fit_Y, fit_M}{Fitted models for Y and M.}
}
\value{
The covariance matrix of all ENC estimates.
}
\description{
Covariance matrix of all ENC estimates
}
