% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R
\name{all_ENCs}
\alias{all_ENCs}
\title{Expected nested counterfactuals for all configurations of \eqn{X} and \eqn{X_M}.}
\usage{
all_ENCs(w, b_Y, theta_Y, b_M, theta_M)
}
\arguments{
\item{w}{Level of covariates, \eqn{W}.}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}
}
\value{
A vector of expected nested counterfactuals for all configurations of \eqn{X} and \eqn{X_M}. Order of output is \code{ENC(1,1), ENC(1,0), ENC(0,1), ENC(0,0)}.
}
\description{
Expected nested counterfactuals for all configurations of \eqn{X} and \eqn{X_M}.
}
