% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R
\name{all_ENCs}
\alias{all_ENCs}
\alias{all_ENCs_theta}
\title{Expected nested counterfactuals for all configurations of \eqn{X} and \eqn{X_M}.}
\usage{
all_ENCs(
  w,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

all_ENCs_theta(w, Theta, which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X"))
}
\arguments{
\item{w}{Level of covariates, \eqn{W}.}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}

\item{which_REs}{Which random effects to include in the calculation. Default is all. Shorthands are available. See details.}

\item{Theta}{A vector of all parameters from both models.}
}
\value{
A vector of expected nested counterfactuals for all configurations of \eqn{X} and \eqn{X_M}. Order of output is \code{ENC(1,1), ENC(1,0), ENC(0,1), ENC(0,0)}.
}
\description{
Expected nested counterfactuals for all configurations of \eqn{X} and \eqn{X_M}.
}
\details{
The following shorthands for random effects are available:
\itemize{
\item "All": All REs
\item "Y.All": All REs for Y
\item "M.All": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
