% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Med_Effs.R
\name{all_MEs}
\alias{all_MEs}
\alias{all_MEs_pars}
\alias{all_MEs_models}
\title{Total, direct and indirect mediation effects on the specified scale(s)}
\usage{
all_MEs_pars(
  scale = c("diff", "rat", "OR"),
  w,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  x_ref = 1,
  x_m_ref = 0,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

all_MEs_models(
  scale = c("diff", "rat", "OR"),
  w,
  fit_Y,
  fit_M,
  x_ref = 1,
  x_m_ref = 0,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)
}
\arguments{
\item{scale}{The scale(s) of the mediation effect. Can be "diff", "rat" or "OR".}

\item{w}{Level of covariates, \eqn{W}.}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}

\item{x_ref, x_m_ref}{Reference values of \eqn{X}, respectively for \eqn{X} and for determining the value of \eqn{M}.}

\item{which_REs}{Which random effects to include in the calculation. Default is all. Shorthands are available. See details.}

\item{fit_Y, fit_M}{Models for \eqn{Y} and \eqn{M}, respectively.}
}
\value{
All mediation effects (total, direct and indirect), on the specified scale(s). Order is total, direct, indirect. Within each effect, order is as specified in \code{scale}. Default is difference, ratio, odds-ratio.
}
\description{
Total, direct and indirect mediation effects on the specified scale(s)
}
\details{
The following shorthands for random effects are available:
\itemize{
\item "All": All REs
\item "Y.All": All REs for Y
\item "M.All": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
