% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R
\name{ENC}
\alias{ENC}
\title{Expected nested counterfactual. Binary Y, binary M.}
\usage{
ENC(x, x_m, w, b_Y, theta_Y, b_M, theta_M)
}
\arguments{
\item{x}{Level of exposure variable, \eqn{X}.}

\item{x_m}{Level of exposure for mediator. I.e. Mediator set to \eqn{M_{x\_m}}.}

\item{w}{Level of covariates, \eqn{W}.}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}
}
\value{
The expected nested counterfactual of \eqn{Y} with \eqn{X = x} and \eqn{M = M_{x\_m}}.
}
\description{
Expected nested counterfactual. Binary Y, binary M.
}
\details{
Contents of \code{b_Y} are \code{(b_Y_0, b_Y_X, b_Y_M, B_Y_W)}. Contents of \code{b_M} are \code{(b_M_0, b_M_X, B_M_W)}.

Contents of \code{theta_Y} are \code{(s_Y_0, cor_Y_0X, cor_Y_0M, s_Y_X, cor_Y_XM, s_Y_M)}. Contents of \code{theta_M} are \code{(s_M_0, cor_M_0X, s_M_X)}.
}
