% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R, R/psi.R
\name{grad_psi}
\alias{grad_psi}
\alias{grad_psi_Y}
\alias{grad_psi_M}
\alias{d1_psi}
\alias{d2_psi}
\title{Gradient of psi used to compute a probability for Y or M}
\usage{
grad_psi_Y(
  m,
  x,
  x_m,
  w,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

grad_psi_M(
  m,
  x,
  x_m,
  w,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

d1_psi(mu, sigma)

d2_psi(mu, sigma)
}
\arguments{
\item{m}{Value of M}

\item{x, x_m}{values of X in the models for Y and M respectively}

\item{w}{Vector of covariates}

\item{b_Y, b_M}{Fixed effects for Y and M models respectively}

\item{theta_Y, theta_M}{SDs and correlations for Y and M models respectively}

\item{which_REs}{Which random effects to include in the calculation. Default is all. Shorthands are available. See details.}

\item{mu}{Mean of the latent Gaussian.}

\item{sigma}{SD of the latent Gaussian.}
}
\value{
Gradient of the function \code{psi(mu, sigma)} with respect to \code{b_Y}, \code{theta_Y}, \code{b_M}, and \code{theta_M}.

The gradient of the function \code{psi(mu, sigma)} with respect to \code{mu} and \code{sigma}.
}
\description{
Gradient of psi used to compute a probability for Y or M

Gradient of the intermediate integral function, psi
}
\details{
The following shorthands for random effects are available:
\itemize{
\item "all": All REs
\item "Y.all": All REs for Y
\item "M.all": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
