% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R, R/psi.R
\name{grad_psi}
\alias{grad_psi}
\alias{grad_psi_Y}
\alias{grad_psi_M}
\alias{d1_psi}
\alias{d2_psi}
\title{Gradient of psi used to compute a probability for Y or M}
\usage{
grad_psi_Y(m, x, x_m, w, b_Y, theta_Y, b_M, theta_M)

grad_psi_M(m, x, x_m, w, b_Y, theta_Y, b_M, theta_M)

d1_psi(mu, sigma)

d2_psi(mu, sigma)
}
\arguments{
\item{m}{Value of M}

\item{x, x_m}{values of X in the models for Y and M respectively}

\item{w}{Vector of covariates}

\item{b_Y, b_M}{Fixed effects for Y and M models respectively}

\item{theta_Y, theta_M}{SDs and correlations for Y and M models respectively}

\item{mu}{Mean of the latent Gaussian.}

\item{sigma}{SD of the latent Gaussian.}
}
\value{
Gradient of the function \code{psi(mu, sigma)} with respect to \code{b_Y}, \code{theta_Y}, \code{b_M}, and \code{theta_M}.

The gradient of the function \code{psi(mu, sigma)} with respect to \code{mu} and \code{sigma}.
}
\description{
Gradient of psi used to compute a probability for Y or M

Gradient of the intermediate integral function, psi
}
