% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_Counterfactual.R
\name{Jacob_ENC}
\alias{Jacob_ENC}
\alias{Jacob_ENC_pars}
\alias{Jacob_ENC_models}
\title{"Jacobian" of expected nested counterfactuals. I.e. Gradients for every combination of \eqn{X, X_M}.}
\usage{
Jacob_ENC_pars(w, b_Y, theta_Y, b_M, theta_M)

Jacob_ENC_models(w, fit_Y, fit_M)
}
\arguments{
\item{w}{Level of covariates, \eqn{W}.}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}

\item{fit_Y, fit_M}{Fitted models for Y and M.}
}
\value{
Gradients for every combination of \eqn{X} and \eqn{X_M}, organized as a 4-by-many matrix. Order of \eqn{(X, X_M)} levels is (1,1), (1,0), (0,1), (0,0).
}
\description{
"Jacobian" of expected nested counterfactuals. I.e. Gradients for every combination of \eqn{X, X_M}.
}
