% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Helpers.R
\name{logit_expit}
\alias{logit_expit}
\alias{logit}
\alias{expit}
\title{Logistic and Inverse-Logistic (expit) transformations}
\usage{
logit(x)

expit(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector of the same length as x.
}
\description{
These functions are useful for transforming probabilities to the real line and back. logit(x) = log(x/(1-x)) and expit(x) = exp(x)/(1+exp(x)).
}
\examples{
logit(c(0.1, 0.2, 0.3))
expit(c(0.1, 0.2, 0.3))

logit(expit(c(0.1, 0.2, 0.3)))
expit(logit(c(0.1, 0.2, 0.3)))
}
