% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Sample_Data.R
\name{make_validation_data}
\alias{make_validation_data}
\title{Generate a dataset for validating methodology}
\usage{
make_validation_data(
  n,
  K,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  num_bin_confounders = 1,
  num_cont_confounders = 1,
  output_list = FALSE,
  return_REs = FALSE,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)
}
\arguments{
\item{n}{Number of observations in each group.}

\item{K}{Number of groups}

\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}

\item{num_bin_confounders, num_cont_confounders}{Number of binary and continuous confounders, respectively.}

\item{output_list}{Should output be formatted as a list with one component per group (of size n-by-5) or a single data.frame of size (Kn)-by-6 with a column labelled `group`?}

\item{return_REs}{Should random effects also be returned? If TRUE, output is a list containing two components: `data` and `REs`.}

\item{which_REs}{Which random effects to include in the calculation. Default is all. See the \href{../vignettes/which_REs.Rmd}{vignette} for more details.}
}
\value{
A simulated dataset, optionally inside a list which also contains random effects.
}
\description{
Generate a validation dataset with `n` observations in each of `K` groups. Variables include one outcome (Y), one mediator (M), one exposure (X), and two confounders (C1 and C2). All variables are binary. Default values are available for regression parameters (fixed effects and covariances of random effects).

Order of variables is `Y`, `M`, `X`, `C1`, `C2`
}
\examples{
n = 20
K = 3
all_reg_pars = make_all_reg_pars()

# Format output as a list
make_validation_data(n, K, all_reg_pars, output_list = TRUE)

# Format output as a data.frame
make_validation_data(n, K, all_reg_pars, output_list = FALSE)
}
