% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Average_ENCs.r
\name{Marginal_ENCs}
\alias{Marginal_ENCs}
\alias{mean_ENC_pars}
\alias{mean_ENC_Theta}
\title{Marginal ENC}
\usage{
mean_ENC_pars(
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  all_confounders,
  confounder_probs,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

mean_ENC_Theta(
  Theta,
  all_confounders,
  confounder_probs,
  which_REs,
  len_par_vecs
)
}
\arguments{
\item{b_Y, b_M}{Coefficient vectors for \eqn{Y}-model and \eqn{M}-model, respectively.}

\item{theta_Y, theta_M}{Covariance parameters of random effects in \eqn{Y}-model and \eqn{M}-model, respectively. See details.}

\item{all_confounders}{A list of all possible values of the confounder variables. Each element is a vector.}

\item{confounder_probs}{A vector of probabilities, with element `i` corresponding to the probability of the `i`th element of `all_confounders`.}

\item{which_REs}{Which random effects to include in the calculation. Default is all. Shorthands are available. See details.}

\item{Theta}{Vector of parameters from both models. Order is b_Y, theta_Y, b_M, theta_M}

\item{len_par_vecs}{Number of entries in each parameter vector. Order is b_Y, theta_Y, b_M, theta_M}
}
\value{
A vector of marginal expected nested counterfactuals (i.e. averaged over the confounder distribution). Order of output is \code{ENC(1,1), ENC(1,0), ENC(0,1), ENC(0,0)}, where the former argument is \eqn{X} and the latter is \eqn{X_M}.
}
\description{
Compute the marginal ENC by averaging conditional ENCs over the provided confounder distribution
}
\details{
The following shorthands for random effects are available:
\itemize{
\item "All": All REs
\item "Y.All": All REs for Y
\item "M.All": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
