% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Med_Effs.R
\name{ME_covariances}
\alias{ME_covariances}
\alias{all_cov_MEs}
\alias{all_covs_MEs_models}
\alias{all_covs_MEs_pars}
\title{Covariance matrix of all mediation effects on various scales.}
\usage{
all_cov_MEs(
  scale = c("diff", "rat", "OR"),
  w,
  fit_Y,
  fit_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

all_covs_MEs_models(
  scale = c("diff", "rat", "OR"),
  w,
  Sigma,
  fit_Y,
  fit_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)

all_covs_MEs_pars(
  scale = c("diff", "rat", "OR"),
  w,
  Sigma,
  b_Y,
  theta_Y,
  b_M,
  theta_M,
  which_REs = c("Y.Int", "Y.X", "Y.M", "M.Int", "M.X")
)
}
\arguments{
\item{scale}{The scale(s) of the mediation effect. Can be "diff", "rat" or "OR".}

\item{w}{Level of covariates, \eqn{W}.}

\item{fit_Y, fit_M}{Fitted models for Y and M.}

\item{which_REs}{Which random effects to include in the calculation. Default is all. See the \href{../vignettes/which_REs.Rmd}{vignette} for more details.}
}
\value{
A covariance matrix for all mediation effects (total, direct and indirect) on the specified scale(s).
}
\description{
Covariance matrix of all mediation effects on various scales.
}
\details{
Note: Uses the \eqn{K}-adjusted covariance matrix, not the asymptotic covariance matrix.

The following shorthands for random effects are available:
\itemize{
\item "All": All REs
\item "Y.All": All REs for Y
\item "M.All": All REs for M
}
Additionally, individual REs can be specified:
\itemize{
\item "Y.Int": Intercept for Y
\item "Y.X": Slope for X in Y
\item "Y.M": Slope for M in Y
\item "M.Int": Intercept for M
\item "M.X": Slope for M
}
}
